fixes a race condition preventing idle() to return as expected
The registered SimpleMessageCountListener is trying to cancel the
current idle()-call. It assumes that calling folder.isOpen() does that.
It will in most cases makes idle() stop by calling noop() internally (to
keep the connection alive) which leads to the current idle()-call being
canceled.

The problem this commits addresses is that the call to noop() only
happens if there were no calls in the last second. There is a check for
that in com.sun.mail.imap.IMAPFolder.keepConnectionAlive. So if a new
message appears less then a second after idle started, we will miss it.